Move theme constants out of ThemeContext
