+++ src/theme/ThemeContext.tsx
@@ -11,14 +11,7 @@ import { onHandleStatusBar } from '@/navigation/onNavigationStateChange';
 
 import { darkModeThemeColors, lightModeThemeColors, type Colors } from '../styles/colors';
 import currentColors from './currentColors';
-
-export const Themes = {
-  DARK: 'dark',
-  LIGHT: 'light',
-  SYSTEM: 'system',
-} as const;
-
-export type ThemesType = (typeof Themes)[keyof typeof Themes];
+import { Themes, type ThemesType } from './themes';
 
 export interface ThemeContextProps {
   colors: Colors;
