+++ src/theme/themes.ts
@@ -1,0 +1,7 @@
+export const Themes = Object.freeze({
+  DARK: 'dark',
+  LIGHT: 'light',
+  SYSTEM: 'system',
+});
+
+export type ThemesType = (typeof Themes)[keyof typeof Themes];
